feat: [IOPLT-1286] Refactor `MessageDetailsStickyFooter` to adopt `FooterActions` (and add dark mode support) (#7230)
## Short description
This PR refactors the `MessageDetailsStickyFooter` to use
`FooterActions` instead of recreating the same structure with simple
buttons as before. Due to the native dark mode support of
`FooterActions`, this change also extends support to this block on the
`MessageDetails` template screen.

## List of changes proposed in this pull request
- Add `mapFooterDataToActions` that handle all the cases by returning
the appropriate `FooterActions` conf object
- Remove `MessageDetailsPaymentButton` (and relative tests) because it's
now included in the `MessageDetailsStickyFooter` component
- Update snapshots

### Preview

https://github.com/user-attachments/assets/814199b7-a00a-48b7-a85f-ac100d640f0e


## How to test
Carefully check all the possible button combinations added in the
following PR:
* https://github.com/pagopa/io-app/pull/5636

---------

Co-authored-by: Andrea <andrea.piai@pagopa.it>