feat(IT Wallet): [SIW-2167,SIW-2476] Events tracking for credentials lifecycle (#7222)
## Short description
This PR introduces Mixpanel tracking events related to the credential
lifecycle.
It also updates the alert copy and adds a CTA button in the bottom sheet
displayed for the Driving License when it's about to expire.

## List of changes proposed in this pull request
- Updated alert logic in `ItwPresentationCredentialStatusAlert` based on
credential status (`jwtExpiring`, `expiring`, `expired`)
- Add tracking events for `tap_banner`, `open_bottom_sheet`, `press_cta`
in `ItwPresentationCredentialStatusAlert`
- For expiring Driving Licenses, a bottom sheet now includes updated
content and a CTA

## How to test

- Verify the correct dispatch of the new tracking events as specified in
Figma, and the updated bottom sheet for expiring alerts when the
credential is a Driving License.

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>