feat: [IOCOM-2532] Universal/App link for SEND dev environment (#7283)
## Short description
This PR adds Universal (iOS) and App (Android) Link support for SEND dev
environment

## List of changes proposed in this pull request
- Associated domain added to Entitlements on iOS 
-
[documentation](https://developer.apple.com/documentation/xcode/supporting-associated-domains#Add-the-associated-domains-entitlement-to-your-app)
- Intent-filter added to Manifest on Android 
- [App Link
documentation](https://developer.android.com/training/app-links#android-app-links)
- [data-element
documentation](https://developer.android.com/guide/topics/manifest/data-element#path)

## How to test
On iOS, run the application on the simulator (or on a real device),
create a reminder (or a message) with the following linked url
`https://cittadini.dev.notifichedigitali.it/io?aar=asd123` and check
that, upon tapping the link, the IO App is opened. Also check that the
base url `https://cittadini.dev.notifichedigitali.it/` does not open the
application but the browser.
In order to test this on Android, the build must be uploaded to the
stores, so that the auto-verification process can happen with the proper
signing certificate. As for this PR, check that the syntax is correct
with respect to the linked documentation and the associated online
files:
- [Asset Links -
Android](https://cittadini.dev.notifichedigitali.it/.well-known/assetlinks.json)
- [App Site Associations -
iOS](https://cittadini.dev.notifichedigitali.it/.well-known/apple-app-site-association)