feat(IT Wallet): [SIW-2814] Implements credential upgrade flow (#7254)
## Short description
This PR introduces the credential upgrade flow. This feature enables
users to update their existing credentials to the new format directly
from their wallet.

## List of changes proposed in this pull request
- Refactored `ItwCredentialIssuanceMachine`:
- Added `mode` parameter to the `select-credential` event to indicate if
the flow is for the issuance, reissuance or upgrade of the credential.
This mode changes how the machine behaves in some states, like if there
is the need to navigate to a certain screen or not.
- Added `EvaluateFlow` state to enhance machine flow readability and
maintenability
- Updated `stroreCredential` action, which now makes sure to delete any
existing credential with the same type before saving the new one
- Added `operationType` parameter to the `obtainCredential` utility
function, which allow to pass the required header for the credential
upgrade

## How to test
- Start with the L3 feature flag disabled. Obtain a new wallet with at
least one credential.
- Restart the app and make sure to have the L3 feature flag enabled.
- Tap the IT Wallet upgrade banner and complete the PID issuance.
- Navigate to your wallet. You should see your old credentials displayed
with an "Upgrade" badge.
- Tap on one of the marked credentials and complete the re-issuance
flow.
- Confirm that the flow completes successfully. Using Reactotron, verify
that the old credential is removed and replaced by the newly obtained
one in the wallet's state.

Additionally check that the async-issuance (from MDL message) and
reissuance (from jwt expired alert message) flow works as expected.

## Demo


https://github.com/user-attachments/assets/59bfbdba-fbe0-42f0-a0ac-250cfd781d6f

---------

Co-authored-by: Cristian Matteu <94987118+ChrisMattew@users.noreply.github.com>
Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>