chore(IT Wallet): [SIW-2765] New status assertion for EAAs (#7227)
> [!WARNING]
> Depends on https://github.com/pagopa/io-react-native-wallet/pull/274

## Short description
This PR introduces the new status assertion for EAAs. To simplify the
upcoming removal of `io-react-native-wallet` double version, **all
status assertions use the new API 1.0**: for this reason **the PR should
be merged for the release that routes all users to the new Issuer
services** and not before.

## List of changes proposed in this pull request
- Re-enabled the status assertion state in the issuance machine (mdoc
credentials are excluded)
- Re-enabled the status assertion fetch at app startup
- Requested the new Issuer Configuration before the status assertion for
legacy credentials with a shared promise to avoid multiple HTTP calls
- Refactored `enrichIssuerResponseError` to be more generic and reused
for status assertion

## How to test
#### IT-Wallet
- Activate IT-Wallet and obtain a credential: inspect the network for a
HTTP call to `/1-0/status`
- Wait for the assertion to expire (or mock it returning true in
`shouldRequestStatusAttestation`): inspect the network for a HTTP call
to `/1-0/status`
#### Documenti su IO (PID 1.0)
- Activate Documenti su IO (L3 flag enabled, but without CIE) and obtain
a credential: inspect the network for a HTTP call to `/1-0/status`
- Wait for the assertion to expire (or mock it returning true in
`shouldRequestStatusAttestation`): inspect the network for a HTTP call
to `/1-0/status`
#### Backward compatibility (eID 0.7)
- Activate Documenti su IO (**L3 flag disabled**) and obtain a
credential v0.7
- Wait for the assertion to expire (or mock it returning true in
`shouldRequestStatusAttestation`): inspect the network for a HTTP call
to `/1-0/status`

---------

Co-authored-by: ChrisMattew <cristian.matteu@spindox.it>
Co-authored-by: Cristian Matteu <94987118+ChrisMattew@users.noreply.github.com>
Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>