chore: [IOBP-1948] Add `intent://` schema inside the payments origin webview whitelist schemas (#7294)
## Short description
This PR makes a small change to the `WalletPaymentWebView.tsx` file by
adding support for the `intent://*` schema to the
`originSchemasWhiteList` array. This update allows the webview to handle
URLs with the `intent://*` schema, which is commonly used for deep
linking in mobile applications on Android.

## List of changes proposed in this pull request
- Added `intent://*` schema to the whitelisted origin schemas;

## How to test
On the dev server, mock any page (e.g., the payment authorization page),
add an Android intent link to an external app, and verify that tapping
it launches the app successfully