feat: [SIW-2842] Enable IT-Wallet for whitelisted users with PID L3 (#7260)
> [!IMPORTANT]
> Depends on #7280 
## Short description
This PR restricts IT-Wallet features (UI, proximity presentation, etc.)
to users who have a PID L3 (1.0) and are included in the whitelist

## List of changes proposed in this pull request
- Upgraded `@pagopa/io-react-native-wallet-v2` version
- Improved `itwLifecycleIsITWalletValidSelector` to check whether the
user is whitelisted
- Improved the condition that determines whether to show the new `mdl`
credential detail screen

## How to test
The only scenario in which IT-Wallet and all its features should be
visible is when the user is whitelisted and has a PID obtained via the
1.0 APIs with an `assurance_level` set to "high" (in short, a PID L3
obtained with CIE + PIN).
In all other cases, we should still display Documents within the IO app.

> [!IMPORTANT]
>Regarding the `mdl` detail view, in order to display the new UI and the
CTA to generate the engagement QR Code for proximity presentation, the
user must meet the requirements to activate IT-Wallet and, in addition,
must have the `mdl` in the new `dc+sd-jwt` format.
>
>**Note:** We're making the assumption based solely on the `dc+sd-jwt`
format because it is the credential used for rendering the detail view.
Furthermore, if that format is present, the `mso_moc` format will also
be available, as both are requested together in the new issuance flow.

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>
Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>