feat: [IOBP-1898] Add show merchants button when webview url present (#7323)
## Short description
This PR introduces a new "Show merchants" secondary action to the IDPay
initiative details screen, provided the `webViewUrl` is included in the
response, allowing users to view participating merchants that redirect
to an external link specified in the backend response.

## List of changes proposed in this pull request
- Added a "Show merchants" (`secondaryCta`) button to the initiative
details footer, which opens the initiative's `webViewUrl` in the browser
if available; otherwise, it displays an error toast.

## How to test
With the dev-server started, verify that a bonus of type `DISCOUNT` with
the `webViewUrl` property filled shows a secondary CTA that redirects to
the provided URL as an external link, opening the device's browser.

## Preview


https://github.com/user-attachments/assets/84c79f6b-ca21-46d0-a9d4-196f2bd253c9