refactor: [SIW-2897] Request `mdoc` credentials only for whitelisted PID L3 (#7291)
## Short description
This PR adds the PID level check during the credential request phase.
The check ensures that formats other than `dc+sd-jwt` are excluded if
the available PID does not meet the required security level (PID L3

## How to test
Using Reactotron or a proxy tool, ensure that the `mdl` in both formats
is requested only if a `PID L3` is available.
If an `eID` is used (by requesting credentials through flow 1.0) or a
`PID L2` is available, the `mdl` must only be requested in the
`dc+sd-jwt` format

---------

Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>