feat:[IOCOM-2544] SEND AAR UAL + pre-login UAL storage (#7328)
## Short description
[DEPENDS ON pagopa/io-dev-api-server/pull/530]

this PR introduces the logic required to both handle an
Universal/AppLink and to store a link of said kind in case the app is
currently closed and can't process it due to remote configurations.

## List of changes proposed in this pull request
- addition of deepLinking reducer to store said link
- addition of logic that opens the SEND flow if the app was in
foreground when a SEND AAR link wakes it up
- required tests

## How to test

using both an android and iOS device, and while running the dev-server
on the branch related to the PR this depends on, make sure that opening
an URL like `https://cittadini.dev.notifichedigitali.it/io?aar=asd123`
(either via another app, like "reminders", or when scanning a QR
containing said URL) has the following behaviour:
- if the app is killed (neither in background nor in foreground), then
it should simply open and run like normal
- if it is in backgound, but not killed , it will go to foreground and
display the SEND AAR flow, even if it requests biometri authentication
before so.

automated tests should also pass

---------

Co-authored-by: Andrea <andrea.piai@pagopa.it>