feat(IT Wallet): [SIW-2904] Implement feature unavailable toast (#7334)
## Short description
This PR introduces a new hook called `useNotAvailableToastGuard`, which
can wrap a function and replace its functionality with a toast
displaying an info alert.

## List of changes proposed in this pull request
- Added `useNotAvailableToastGuard`
- Refactored the header inside `ItwPresentationDetailsScreenBase`

![photo_2025-09-10 16 24
29](https://github.com/user-attachments/assets/6d6b849f-4cc5-4b2c-9910-14ba831a6dab)

## How to test
Activate ITW and verify that each wrapped CTA shows the alert:
- The header "?" icon  
- The credential footer CTA "Qualcosa non torna"  
- The "Powered by" CTA inside the PID L3  

Also verify that no regressions are present in the **Documenti su IO**
version of these screens.

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>