fix: [IOPLT-1336] Remove footer gradient from `ItwPresentationDetailsScreenBase` + Fix barcode (#7347)
## Short description
This PR removes the footer gradient from
`ItwPresentationDetailsScreenBase`, which is mainly used to display the
fiscal code to citizens. This white gradient is also visible in dark
mode, creating a stark contrast that is a visual mistake.

## List of changes proposed in this pull request
- Refactor `ItwPresentationDetailsScreenBase`
- Fix the fiscal code barcode by using a white rectangle, even in dark
mode version

### Preview
I can't post a screenshot because you can't take screenshots in that
page.

## How to test
N/A

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>