chore(IT Wallet): [SIW-2961] Store whether IT-Wallet can be activated with the simplified flow (#7337)
## Short description
This PR stores a boolean value indicating whether the user will be able
to activate IT-Wallet with the simplified flow.

During the initial testing period users that are **not whitelisted**
might obtain an L3 PID; however they will still see "Documenti su IO"
until the public release of IT-Wallet, and then will be able to upgrade
using a simplified flow. Therefore it is necessary to distinguish
between
- Whitelisted users with L3 PID (they do not need to upgrade)
- Non-whitelisted users with L3 PID (they need to upgrade but with a
simplified flow)

Simply checking whether the PID is a valid ITW credential is not enough.

## List of changes proposed in this pull request
- Persist L3 + whitelist information when storing the PID

## How to test
Activate Documenti su IO/IT-Wallet with the whitelist on/off. Ensure the
new property is true only with the combination PID L3 + whitelist off.