feat(IT Wallet): [SIW-2850] Introduce an intermediate screen after a 5-second wait (#7324)
## Short description
This PR includes the introduction of an intermediate screen that is
displayed if the trust check takes longer than 5 seconds in the
activation, document issuance, and remote presentation flows.

## List of changes proposed in this pull request
- Added a 5-second timeout (`after`) in the
`TrustFederationVerification` for `credential`
- Added a 5-second timeout (`after`) in the
`TrustFederationVerification` for `eid`
- Implemented loading state handling in `ItwIpzsPrivacyScreen` and
`ItwL2IdentificationModeSelectionScreen`
- Added a `timeout` key to the Italian translations
- Introduced timeout handling in `ItwRemoteRequestValidationScreen`
- Updated `ItwRemoteLoadingScreen` to accept `testID` as an optional
prop (for testing purposes)
- Added tests for all newly introduced extended-loading timeouts

## How to test
- **credential**: add a document 
- **eid**: activate `Documenti su IO` or ITWallet
- **presentation**: perform a remote presentation of a credential

Since this feature depends on the user’s connection and the screen won’t
always appear, tests has been added to ensure the screen does not open
before the 5-second threshold and opens exactly once after 5 seconds.
To test: run `yarn jest [relative path of machine.test.ts]`

## Screen Preview

For testing purposes, the timeout is set to 1 second for the first two
cases, while for the third one it is set to 5 seconds.

Credential:


https://github.com/user-attachments/assets/c8981f9e-6b02-4d05-acd7-5273c1f00668


Eid:


https://github.com/user-attachments/assets/2920b76d-7c5b-42da-9bb3-7f60bdcd874d


Presentation:


https://github.com/user-attachments/assets/e1af3a67-42eb-4a83-aad5-c75983b41696

---------

Co-authored-by: MonicaRungi-w <monica.rungi@itauros.com>