fix: [IOBP-1948] Android payment webview intent linking (#7369)
## Short description
This PR makes a small change to how URLs are handled in the
`WalletPaymentWebView` component intent linking. The update ensures that
intent URLs are properly decoded before being opened, which helps
prevent issues with encoded characters in URLs.

## List of changes proposed in this pull request
* WebView intent handling: Updated the call to `Linking.openURL` to use
`decodeURIComponent` on the intent URL, ensuring that any encoded
characters are correctly interpreted before opening the URL.

## Preview
The following video shows that the `Poste Italiane` app is opened when
the user taps on the "Autorizza in app" CTA inside the webview.


https://github.com/user-attachments/assets/ecba3b6c-5b8a-4a66-bb29-1da44ada9443