feat(IT Wallet): [SIW-2945] Add IT Wallet settings screen (#7355)
## Short description
This PR adds the settings screen for the IT Wallet feature, only visible
to the whitelisted users.
The settings screen allows to view the current status of the IT wallet
and shows a CTA to start the revocation flow. If the user does not have
the IT Wallet already active, the screen shows a CTA for its activation.

## List of changes proposed in this pull request
- Added IT-Wallet menu item in the settings screen.
- Added `isHidden` prop to `ProfileNavListItem` to allow programmatic
hide/show for settings menu items
- Added `ItwSettingsScreen`
- Added tests

## How to test
- Verify that the menu item is only visible when the L3 whitelist flag
is enabled.
- Verify that when you navigate to the settings screen **without** a
valid IT Wallet, it shows a CTA to activate it
- Verify that when you navigate to the settings screen **with** a valid
IT Wallet, it shows the activation status and a CTA to revoke it
- Verify that tapping on the link shows a toast

## Preview

| Without IT Wallet | With IT Wallet |
| --- | --- |
| <img width="250"
src="https://github.com/user-attachments/assets/72225b5b-870d-423f-a748-b45659176afa"
/> | <img width="250"
src="https://github.com/user-attachments/assets/ba622a79-f2a5-48a6-9658-834595315a6c"
/> |