feat(IT Wallet): [SIW-2759] Add simplified ITW activation flow (#7362)
## Short description
This PR enables the simplified activation flow for users that have a
valid L3 PID obtained while not being whitelisted. When they're added to
the whitelist they should see the IT-Wallet banner and simply accept the
TOS to continue, without re-authenticating.

## List of changes proposed in this pull request
- Added guard to check for the simplified activation requirements
- Added new state to evaluate the simplified activation

## How to test
1. Disable the L3 flag
2. Activate the wallet with CIE+PIN and ensure the property
`preferences.isItwSimplifiedActivationRequired` is true
3. Restart the app to be in the whitelist
4. Ensure the banner to activate IT-Wallet is visible and activate the
wallet
5. You should not need to re-authenticate after accepting the
TOS/privacy policy

Repeat the steps with an additional credential to upgrade: you should
skip the authentication step and only see the upgrade step.

Ensure no regressions were introduced: activate the wallet with
different combinations of whitelist and L2/L3. The simplified flow must
be **enabled only if the PID was obtained with CIE+PIN with the
whitelist off**.

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>