fix(IT Wallet): [SIW-3092] Wait for integrity service before getting the wallet attestation in remote presentation (#7414)
## Short description
This PR fixes a potential bug that arises in the same-device remote
presentation flow, when the app is launched to the presentation screens
and it does not have enough time to warm up the integrity service,
invoked before requesting a wallet attestation.

## List of changes proposed in this pull request
- Added `pollForStoreValue` in the `getWalletAttestation` actor


## How to test
The bug is not easy to reproduce, as it is a race condition that does
not always happen.

To test it, it is necessary to try several same-device presentations
with an expired wallet attestation. Before the fix, the presentation may
fail with the error `PREPARE_NOT_CALLED`.

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>