chore(IT Wallet): [SIW-3048] Add new Documenti su IO activation fallback screen (#7409)
## Short description
This PR introduces the revised Documenti su IO activation screen,
displayed when the user does not possess a CIE card.

It is only available in the playground, as the **current screen must not
be changed yet**.

## List of changes proposed in this pull request
- Renamed the current screen to `ItwLegacyDiscoveryInfoComponent`
- Created `ItwDiscoveryInfoComponent`
- Modified IOMarkdown rules to support small paragraphs and links

## How to test
Navigate to ITW Playground and see the new screen. Ensure other
instances of IOMarkdown renders correctly.



https://github.com/user-attachments/assets/cb76d01c-ebfb-4980-84d7-8c9a15700976