fix: [IOBP-2038] Set deviceId to the CdC url only if mixpanel enabled (#7436)
## Short description
This pull request updates the `CdcBonusRequestInformationTos` screen to
ensure that the device ID is only included in the request URL if
Mixpanel tracking is enabled.

## List of changes proposed in this pull request
* Added a check for Mixpanel enablement by selecting `isMixpanelEnabled`
from persisted preferences.
* Updated the logic to only append the device ID to the URL if both
`includeDeviceId` is true in the config and Mixpanel is enabled
(`ts/features/bonus/cdc/activation/screens/CdcBonusRequestInformationTos.tsx`).

## How to test
Open the IO app and disable tracking. Then, verify through Mixpanel
that, when the user starts the CdC request, the webview does not track
events for that deviceId.