fix(IT Wallet): [SIW-3079,SIW-3109] Improve claims render during Proximity presentation online/offline (#7429)
## Short description
This PR improves the claims parsing during the Proximity presentation
## List of changes proposed in this pull request
- Ordered the claims by the EC claims order
- Implemented a new decoder for the `driving_privileges` claim specific
to the mdoc credential
- Bumped `@pagopa/io-react-native-iso18013` to `0.6.0` to fix Android
offline presentation

## How to test
After obtaining an ITW mDL, start a Proximity presentation and verify
that every claim respects the EC claims order; moreover, verify that no
undefined claims are present. Test everything online/offline on both iOS
and Android

> [!IMPORTANT]
> Since this PR fixes the Offline Proximity presentation on Android, you
must test it before approving this PR. This is a crucial fix that we
have to test.

---------

Co-authored-by: Gianluca Spada <gianluca.spada@spindox.it>