feat: [IOCOM-2550] Send AAR attachment download (#7430)
⚠️ This PR depends on https://github.com/pagopa/io-app/pull/7432 ⚠️

## Short description
This PR add downloading of SEND AAR's attachments.

## List of changes proposed in this pull request
- New saga that retrieves the download metadata and then download the
file
- Such saga is run based on the third party message type from which the
attachment is requested
- Lollipop keyinfo data has been refactored, in order to be shared from
the startup saga

## How to test
Using the io-dev-api-server, change the following files and lines:
- `ts/features/messages/hooks/useAttachmentDownload.tsx` row 109
```
downloadAttachment.request({
  attachment: {
    ...attachment,
    url: "REPLACEME" as NonEmptyString
  },
  messageId: "00000000000000000000000006",
  skipMixpanelTrackingOnFailure: isPN,
  serviceId
})
```
- `ts/features/messages/saga/handleDownloadAttachment.ts` row 197
```
  ephemeralAARThirdPartyMessage: true,
```

and replace `REPLACEME` with the url of an attachment that belongs to
the send notification with ID `SEND-SEND-SEND-000002-A-0` (you can
retrieve it by calling
`http://localhost:3000/api/v1/send/aar/notifications/SEND-SEND-SEND-000002-A-0?isTest=false`
with cUrl/Bruno/Postman/...).

Select the SEND message with title `Nuovo avviso di ingiunzione` and try
to download one of it's attachments

---------

Co-authored-by: Martino Cesari Tomba <60693085+forrest57@users.noreply.github.com>