feat(IT Wallet): [SIW-3125] Add eID reissuance qualtrics survey (#7456)
## Short description
This PR adds a new reissuance feedback banner and bottom sheet alongside
a new Qualtrics survey in specific ITW actions.
## List of changes proposed in this pull request
- Added a new feedback banner that shows when an eID reissuance is
completed
- Added a new hook to present a bottom sheet with a Qualtrics survey
(optional)

## How to test
To test this PR, you have to mock your eID expiration date:
- Verify that in the starting eID expired screen (add a credential with
an expired eID), when pressing on `Annulla`, the Qualtrics bottom sheet
opens. Verify that you can close it and navigate back, and confirm that
the survey button functions properly.
- Verify that whenever you find a native dismissal dialog and you are in
reissuing, when stopping the flow and going back to the wallet home
screen, the survey bottom sheet opens up.
- Verify that, after a successful reissuance of the eID, you can see the
Feedback banner, you can click on it, and you can dismiss it.
Verify that no regressions or issues are present in the issuance flow.