refactor(IT Wallet): [SIW-2879] Rename status attestation to status assertion (#7431)
## Short description
This PR renames all occurrences of credentials status attestation to
status assertion, to be consistent with [IT-Wallet specifications
1.0](https://italia.github.io/eid-wallet-it-docs/versione-corrente/en/credential-revocation.html#oauth-status-assertions).

## List of changes proposed in this pull request
- Renamed status attestation to status assertion
- Added migration with tests

## How to test

#### Test the migration
Get some credentials on the `master` branch, then switch to this branch.
Ensure `storedStatusAttestation` in the stored credential was
successfully renamed to `storedStatusAssertion` containing
```json
{
  "credentialStatus": "",
  "statusAssertion": "",
  "parsedStatusAssertion": {}
}
```

#### Test the assertion fetch
Request new credentials and ensure the assertion is fetched correctly
and stored in the renamed key.