feat: [IOCOM-2653] Change text on the Alert of the timeline of the SEND bottom sheet (#7460)
## Short description
This PR changes the text displayed on the Alert in the bottom sheet of
the SEND's timeline.

<img width="357" height="530" alt="Screenshot 2025-10-09 at 15 01 07"
src="https://github.com/user-attachments/assets/b24276e0-ee81-4192-b81e-4f1dcd61884d"
/>

## List of changes proposed in this pull request
- New copy for it and en translations
- Removed bad carriage return character in italian translations

## How to test
Using the io-dev-api-server, select a SEND message and check that the
Alert in the timeline has the new text.

Co-authored-by: Martino Cesari Tomba <60693085+forrest57@users.noreply.github.com>