feat(IT Wallet): [SIW-3137] Tracking of presentation remote event (#7461)
## Short description
This PR introduces tracking for the new event `ITW_REMOTE_START` and the
new code `ITW presentazione remota` within the `QRCODE_SCAN_SUCCESS`
event.

## List of changes proposed in this pull request
- Added new event: `ITW_REMOTE_START`
- Added new code: `ITW presentazione remota`
- Fix tests

## How to test
Refer to
[Airtable](https://airtable.com/appfuAAW2S44jfutx/tbldC28jZYV4YoEsh/viwop2sOY1gIz5Sbk/recoYQvimkmxxfIrb?blocks=hide)
Verify that the remote presentation event (`ITW_REMOTE_START`) is
triggered correctly when the QR code is scanned.
Ensure that the `code` property for the `QRCODE_SCAN_SUCCESS` event is
populated correctly.

---------

Co-authored-by: MonicaRungi-EY <monica.rungi@itauros.com>