chore: [IOBP-2096] Handle IDPay initiative details error screen (#7512)
## Short description
This pull request enhances error handling for the IDPay initiative
details screen. It adds the operation result screen UI to display a
dedicated error screen when initiative details API requests are in
error.

## List of changes proposed in this pull request
* Implemented a dedicated error screen using
`OperationResultScreenContent` that appears when initiative details fail
to load, hiding the header and providing a "Back" action.
* Added a new Mixpanel event, `trackIDPayDetailError`, to track when
users encounter errors on the initiative details screen.
* Triggered the new error tracking event on first render if initiative
details fail to load.

## How to test
- Simulate an error for the IDPay detail initiative screen and try to
open the initiative detail, checking that the correct error screen
displays

## Preview

https://github.com/user-attachments/assets/2b120608-71c0-4d2d-9870-111dcb8d1b3a