feat: [IOPID-3514] improvements on analytics tracking  (#7520)
## Short description
This pull request introduces improvements to analytics tracking and
error reporting, particularly for session timeouts and ingress API
failures. The changes enhance the granularity of tracked events and
provide more detailed context for errors, which should help with
debugging and analysis.

## List of changes proposed in this pull request
### Analytics and event tracking improvements

* Updated the event name for post-PIN login session timeout tracking to
`"LOGIN_SESSION_TIMEOUT_POST_PIN_V2"` in
`trackLoginSessionTimeoutPostPin` in order to track post-PIN session
timeout when the error is specifically a `"transient-error"`, improving
accuracy of session timeout analytics.

### Ingress API failure diagnostics

* Enhanced `trackIngressTimeout` to accept an `API_failure` parameter,
enabling tracking of which API calls failed during ingress timeouts for
better diagnostics.
* Added logic in `IngressScreen.tsx` to build a detailed `API_failure`
string based on the state of connectivity, session, and profile loading,
and pass it to `trackIngressTimeout` for more informative error
tracking.


https://github.com/user-attachments/assets/1f51d023-f039-4d04-aaad-802ceca0c2ff


## How to test
1. Run the application 
2. using proxyman add a timeout > 8000ms (or error 500) on `/ping`
`/profile` `/session` API